refactor: move create icon to assets directory
